shooter/requirements.txt
-r ../requirements.txt
pyaml==15.02.1
cliff==1.9.0
nose==1.3.4
pbr==0.10.7
pyzmq==14.5.0
six==1.9.0
coverage==4.0a5
pyparsing==2.0.3
stevedore==1.2.0
